Kernel: Take some baby steps towards x86_64
Make more of the kernel compile in 64-bit mode, and make some things
pointer-size-agnostic (by using FlatPtr.)

There's a lot of work to do here before the kernel will even compile.

+++ Kernel/API/Syscall.h
@@ -318,7 +318,7 @@ struct SC_futex_params {
     u32 val;
     union {
         const timespec* timeout;
-        u32 val2;
+        uintptr_t val2;
     };
     u32* userspace_address2;
     u32 val3;
